rename modal components with form: "xxxForm" -> "xxxAsForm"
