+++ admin-app/frontend/src/components/dataModal/survey/SurveyModal.vue
@@ -1,9 +1,9 @@
 <template>
   <SurveyModalFirstCardRead v-if="completionIndex === 1 && read" />
-  <SurveyModalFirstCardForm v-else-if="completionIndex === 1 && !read" />
+  <SurveyModalFirstCardAsForm v-else-if="completionIndex === 1 && !read" />
 
   <SurveyModalQuestionRead v-else-if="completionIndex === 2 && read" />
-  <SurveyModalQuestionForm v-else-if="completionIndex === 2 && !read" />
+  <SurveyModalQuestionAsForm v-else-if="completionIndex === 2 && !read" />
 
   <!-- This step is available only in create mode -->
   <SurveyModalFinalize v-else-if="completionIndex === 3" />
@@ -14,10 +14,10 @@ import { mapGetters } from 'vuex';
 import { modalModesDict } from '../../../lib/constants';
 
 import SurveyModalFirstCardRead from './firstCard/SurveyModalFirstCardRead.vue';
-import SurveyModalFirstCardForm from './firstCard/SurveyModalFirstCardForm.vue';
+import SurveyModalFirstCardAsForm from './firstCard/SurveyModalFirstCardAsForm.vue';
 
 import SurveyModalQuestionRead from './question/SurveyModalQuestionRead.vue';
-import SurveyModalQuestionForm from './question/SurveyModalQuestionForm.vue';
+import SurveyModalQuestionAsForm from './question/SurveyModalQuestionAsForm.vue';
 
 import SurveyModalFinalize from './SurveyModalFinalize.vue';
 
@@ -25,9 +25,9 @@ export default {
   name: 'SurveyModal',
   components: {
     SurveyModalFirstCardRead,
-    SurveyModalFirstCardForm,
+    SurveyModalFirstCardAsForm,
     SurveyModalQuestionRead,
-    SurveyModalQuestionForm,
+    SurveyModalQuestionAsForm,
     SurveyModalFinalize,
   },
   computed: {
